backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
torch==2.4.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
faiss-cpu==1.8.0
tqdm==4.66.4
requests==2.32.3
scipy==1.13.1